refactor conversations routes code to be more modular
